% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_spikes_bad}
\alias{df_spikes_bad}
\title{Example data set with bad variable names illustrating spike package}
\format{
A data frame in wide format with one row per participant and 17 columns:
\describe{
\item{id}{Unique identifyer}
\item{phq_1 - phq_15}{15 repeated measures of PHQ-9 scores}
\item{spike}{variable indicating whether there participant}
}
}
\usage{
df_spikes_bad
}
\description{
Example data set with bad variable names illustrating spike package
}
\keyword{datasets}
