% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_vars.R
\name{rename_spike_vars}
\alias{rename_spike_vars}
\title{Rename variables}
\usage{
rename_spike_vars(data, repeated_measures)
}
\arguments{
\item{data}{Dataset in wide format}

\item{repeated_measures}{List of repeated measures}
}
\value{
Dataset in wide format with renamed variables
}
\description{
Rename variables
}
