% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_spike.R
\name{identidy_spikes}
\alias{identidy_spikes}
\title{Identify symptom spikes}
\usage{
identidy_spikes(
  data,
  id_var,
  repeated_measures,
  cutoff_deterioration,
  cutoff_improvement,
  max_spike_duration,
  min_between_spike_interval,
  tx_phases
)
}
\arguments{
\item{data}{Data in wide format with one row for each participant and one column for each repeated measure}

\item{id_var}{Variable that identifies each participant}

\item{repeated_measures}{List of repeated measures}

\item{cutoff_deterioration}{Minimum number of symptom decrease that would be considered as a spike.}

\item{cutoff_improvement}{Minimum symptom improvement after symptom decrease to be considered a spike.}

\item{max_spike_duration}{Maximum number of repeated measures between symptom deterioration and improvement that would still be considered a spike.}

\item{min_between_spike_interval}{Minimum number of sessions between multiple spikes experiences by the same participant}

\item{tx_phases}{NOT SUPPORTED AT THE MOMENT. List specifying treatment phases.}
}
\value{

}
\description{
Identify symptom spikes
}
